debugging xdebug install

+++ recipes/default.rb
@@ -44,6 +44,7 @@ end
 
 # install composer
 execute "install_composer" do
+  not_if File.exists? "/usr/local/bin/composer"
   command "curl -s https://getcomposer.org/installer | php && mv composer.phar /usr/local/bin/composer"
   action :run
 end
@@ -58,8 +59,8 @@ bash "install_xdebug" do
   code <<-EOH
     wget http://xdebug.org/files/xdebug-#{node[:php][:xdebug_version]}.tgz -O /usr/src/xdebug-#{node[:php][:xdebug_version]}.tgz
     tar -zxf xdebug-#{node[:php][:xdebug_version]}.tgz
-    (cd xdebug-#{node[:php][:xdebug_version]}/ && phpize && ./configure && make)
-    cd /usr/src/xdebug-#{node[:php][:xdebug_version]} && phpize | grep Zend\ Module | sed 's/[^0-9.]*\([0-9.]*\).*/\1/' > phpize_version.txt
+    (cd xdebug-#{node[:php][:xdebug_version]}/ && phpize | grep Zend\ Module | sed 's/[^0-9.]*\([0-9.]*\).*/\1/' > phpize_version.txt)
+    (./configure && make)
     cp /usr/src/xdebug-#{node[:php][:xdebug_version]}/modules/xdebug.so /usr/lib/php5/`cat /usr/src/xdebug-#{node[:php][:xdebug_version]}/phpize_version.txt`/
   EOH
